requirements.txt
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
asttokens==3.0.0
async-timeout==5.0.1
attrs==21.2.0
azure-core==1.35.0
azure-storage-blob==12.26.0
backcall==0.2.0
bcrypt==4.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.9.0
Brotli==1.0.9
cachelib==0.13.0
certifi==2025.6.15
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.2
click==8.2.1
comm==0.2.2
command-not-found==0.3
cryptography==45.0.4
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.8.14
decorator==5.2.1
distro==1.9.0
distro-info==1.1+ubuntu0.1
eval_type_backport==0.2.0
exceptiongroup==1.3.0
executing==2.2.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==5.0.1
Flask-DebugToolbar==0.16.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fonttools==4.29.1
fs==2.4.12
gast==0.5.2
greenlet==3.1.1
gunicorn==23.0.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.20.2
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
importlib-metadata==4.6.4
ipykernel==6.29.5
ipython==8.37.0
ipython-genutils==0.2.0
isodate==0.7.2
itsdangerous==2.2.0
jedi==0.19.2
jeepney==0.7.1
Jinja2==3.1.4
jiter==0.10.0
jsonpath-python==1.0.6
jupyter_client==8.6.3
jupyter_core==5.8.1
keyring==23.5.0
kiwisolver==1.3.2
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lxml==4.8.0
lz4==3.1.3+dfsg
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.1
matplotlib-inline==0.1.7
mcp==1.6.0
mdurl==0.1.2
mistralai==1.9.2
more-itertools==8.10.0
mpmath==0.0.0
msgspec==0.19.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
openai==1.75.0
packaging==25.0
parso==0.8.4
pdfminer.six==20250506
pdfplumber==0.11.7
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
platformdirs==4.3.8
ply==3.11
prompt_toolkit==3.0.51
psutil==7.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
PyGObject==3.42.1
PyJWT==2.3.0
pyOpenSSL==24.2.1
pyparsing==2.4.7
pypdfium2==4.30.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pythran==0.10.0
pytz==2022.1
PyYAML==5.4.1
pyzmq==27.0.0
redis==5.2.1
requests==2.32.3
rich==14.0.0
scipy==1.8.0
SecretStorage==3.3.1
shellingham==1.5.4
simplegeneric==0.8.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.3.1
SQLAlchemy==2.0.35
sse-starlette==2.3.6
stack-data==0.6.3
starlette==0.47.1
stripe==11.2.0
sympy==1.9
systemd-python==234
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
ubuntu-advantage-tools==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==2.2.3
uvicorn==0.34.3
wadllib==1.3.6
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.1.3
WTForms==3.1.2
zipp==1.0.0
